% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_uml.R
\name{add_uml}
\alias{add_uml}
\title{Paste the workflow figure built with the Unified Modeling Language}
\usage{
add_uml(open.r.snippets = TRUE)
}
\arguments{
\item{open.r.snippets}{Logical, by default \code{TRUE}}
}
\value{
If \code{open.r.snippets = FALSE}, the result is the text including the UML text.
If \code{open.r.snippets = TRUE}, open the local R snippet file.
}
\description{
Paste the workflow figure built with the Unified Modeling Language
}
\examples{
add_uml(open.r.snippets = FALSE)
}
